+++ src/Voting.Lib.DmDoc/Models/BrickContent.cs
@@ -1,0 +1,15 @@
+﻿// (c) Copyright 2022 by Abraxas Informatik AG
+// For license information see LICENSE file
+
+namespace Voting.Lib.DmDoc.Models;
+
+/// <summary>
+/// A DmDoc brick content.
+/// </summary>
+public class BrickContent
+{
+    /// <summary>
+    /// Gets or sets the brick content ID.
+    /// </summary>
+    public int Id { get; set; }
+}