+++ src/Voting.Lib.DmDoc/Models/Internal/BrickContentEditResponse.cs
@@ -1,0 +1,9 @@
+﻿// (c) Copyright 2022 by Abraxas Informatik AG
+// For license information see LICENSE file
+
+namespace Voting.Lib.DmDoc.Models.Internal;
+
+internal class BrickContentEditResponse
+{
+    public string EditorUrlV3 { get; set; } = string.Empty;
+}