+++ src/Voting.Lib.DmDoc/Models/Internal/UpdateBrickContentResponse.cs
@@ -1,0 +1,11 @@
+﻿// (c) Copyright 2022 by Abraxas Informatik AG
+// For license information see LICENSE file
+
+namespace Voting.Lib.DmDoc.Models.Internal;
+
+internal class UpdateBrickContentResponse
+{
+    public int Id { get; set; }
+
+    public int BrickId { get; set; }
+}